+++ packages/utils/src/isConstant.ts
@@ -9,5 +9,5 @@ import { RJSFSchema } from './types';
  * @returns - True if the `schema` has a single constant value, false otherwise
  */
 export default function isConstant(schema: RJSFSchema) {
-  return (Array.isArray(schema.enum) && schema.enum.length === 1) || schema.hasOwnProperty(CONST_KEY);
+  return (Array.isArray(schema.enum) && schema.enum.length === 1) || CONST_KEY in schema;
 }
