+++ packages/utils/src/mergeSchemas.ts
@@ -2,7 +2,7 @@ import union from 'lodash/union';
 
 import getSchemaType from './getSchemaType';
 import isObject from './isObject';
-import { GenericObjectType } from 'types';
+import { GenericObjectType } from './types';
 
 /** Recursively merge deeply nested schemas. The difference between mergeSchemas and mergeObjects is that mergeSchemas
  * only concats arrays for values under the 'required' keyword, and when it does, it doesn't include duplicate values.
@@ -16,7 +16,7 @@ export default function mergeSchemas(obj1: GenericObjectType, obj2: GenericObjec
   return Object.keys(obj2).reduce((acc, key) => {
     const left = obj1 ? obj1[key] : {},
       right = obj2[key];
-    if (obj1 && obj1.hasOwnProperty(key) && isObject(right)) {
+    if (obj1 && key in obj1 && isObject(right)) {
       acc[key] = mergeSchemas(left, right);
     } else if (
       obj1 &&
