+++ packages/utils/src/schema/retrieveSchema.ts
@@ -1,13 +1,20 @@
 import get from 'lodash/get';
+import set from 'lodash/set';
 import mergeAllOf from 'json-schema-merge-allof';
 
-import { ADDITIONAL_PROPERTIES_KEY, ALL_OF_KEY, DEPENDENCIES_KEY, REF_KEY } from '../constants';
+import {
+  ADDITIONAL_PROPERTIES_KEY,
+  ADDITIONAL_PROPERTY_FLAG,
+  ALL_OF_KEY,
+  DEPENDENCIES_KEY,
+  REF_KEY
+} from '../constants';
 import findSchemaDefinition, { splitKeyElementFromObject } from '../findSchemaDefinition';
+import guessType from '../guessType';
 import isObject from '../isObject';
 import mergeSchemas from '../mergeSchemas';
 import { GenericObjectType, RJSFSchema, RJSFSchemaDefinition, ValidatorType } from '../types';
 import getMatchingOption from './getMatchingOption';
-import stubExistingAdditionalProperties from './stubExistingAdditionalProperties';
 
 /** Resolves a conditional block (if/else/then) by removing the condition and merging the appropriate conditional branch
  * with the rest of the schema
@@ -58,14 +65,14 @@ export function resolveCondition<T = any>(
 export function resolveSchema<T = any>(
   validator: ValidatorType, schema: RJSFSchema, rootSchema: RJSFSchema = {}, formData?: T
 ): RJSFSchema {
-  if (schema.hasOwnProperty(REF_KEY)) {
+  if (REF_KEY in schema) {
     return resolveReference<T>(validator, schema, rootSchema, formData);
   }
-  if (schema.hasOwnProperty(DEPENDENCIES_KEY)) {
+  if (DEPENDENCIES_KEY in schema) {
     const resolvedSchema = resolveDependencies<T>(validator, schema, rootSchema, formData);
     return retrieveSchema<T>(validator, resolvedSchema, rootSchema, formData);
   }
-  if (schema.hasOwnProperty(ALL_OF_KEY)) {
+  if (ALL_OF_KEY in schema) {
     return {
       ...schema,
       allOf: schema.allOf!.map((allOfSubschema) =>
@@ -101,6 +108,61 @@ export function resolveReference<T = any>(
   );
 }
 
+/** Creates new 'properties' items for each key in the `formData`
+ *
+ * @param validator - An implementation of the `ValidatorType` interface that will be used when necessary
+ * @param theSchema - The schema for which the existing additional properties is desired
+ * @param [rootSchema] - The root schema, used to primarily to look up `$ref`s * @param validator
+ * @param [aFormData] - The current formData, if any, to assist retrieving a schema
+ * @returns - The updated schema with additional properties stubbed
+ */
+export function stubExistingAdditionalProperties<T = any>(
+  validator: ValidatorType,
+  theSchema: RJSFSchema,
+  rootSchema?: RJSFSchema,
+  aFormData?: T
+): RJSFSchema {
+  // Clone the schema so we don't ruin the consumer's original
+  const schema = {
+    ...theSchema,
+    properties: { ...theSchema.properties },
+  };
+
+  // make sure formData is an object
+  const formData: GenericObjectType = aFormData && isObject(aFormData) ? aFormData : {};
+  Object.keys(formData).forEach((key) => {
+    if (key in schema.properties) {
+      // No need to stub, our schema already has the property
+      return;
+    }
+
+    let additionalProperties: RJSFSchema = {};
+    if (typeof schema.additionalProperties !== 'boolean') {
+      if (REF_KEY in schema.additionalProperties!) {
+        additionalProperties = retrieveSchema<T>(
+          validator,
+          { $ref: get(schema.additionalProperties, [REF_KEY]) },
+          rootSchema,
+          formData as T
+        );
+      } else if ('type' in schema.additionalProperties!) {
+        additionalProperties = { ...schema.additionalProperties };
+      } else {
+        additionalProperties = { type: guessType(get(formData, [key])) };
+      }
+    } else {
+      additionalProperties = { type: guessType(get(formData, [key])) };
+    }
+
+    // The type of our new key should match the additionalProperties value;
+    schema.properties[key] = additionalProperties;
+    // Set our additional property flag so we know it was dynamically added
+    set(schema.properties, [key, ADDITIONAL_PROPERTY_FLAG], true);
+  });
+
+  return schema;
+}
+
 /** Retrieves an expanded schema that has had all of its conditions, additional properties, references and dependencies
  * resolved and merged into the `schema` given a `validator`, `rootSchema` and `rawFormData` that is used to do the
  * potentially recursive resolution.
@@ -119,7 +181,7 @@ export default function retrieveSchema<T = any>(
   }
   let resolvedSchema = resolveSchema<T>(validator, schema, rootSchema, rawFormData);
 
-  if (schema.hasOwnProperty('if')) {
+  if ('if' in schema) {
     return resolveCondition<T>(validator, schema, rootSchema, rawFormData as T);
   }
 
@@ -165,7 +227,7 @@ export default function retrieveSchema<T = any>(
     }
   }
   const hasAdditionalProperties =
-    resolvedSchema.hasOwnProperty(ADDITIONAL_PROPERTIES_KEY) &&
+    ADDITIONAL_PROPERTIES_KEY in resolvedSchema &&
     resolvedSchema.additionalProperties !== false;
   if (hasAdditionalProperties) {
     return stubExistingAdditionalProperties<T>(
@@ -309,7 +371,7 @@ export function withDependentSchema<T>(
   }
   // Resolve $refs inside oneOf.
   const resolvedOneOf = oneOf.map((subschema) => {
-    if (typeof subschema === 'boolean' || !subschema.hasOwnProperty(REF_KEY)) {
+    if (typeof subschema === 'boolean' || !(REF_KEY in subschema)) {
       return subschema;
     }
     return resolveReference<T>(validator, subschema as RJSFSchema, rootSchema, formData);
