+++ packages/utils/src/toConstant.ts
@@ -5,7 +5,7 @@ export default function toConstant(schema: RJSFSchema) {
   if (Array.isArray(schema.enum) && schema.enum.length === 1) {
     return schema.enum[0];
   }
-  if (schema.hasOwnProperty(CONST_KEY)) {
+  if (CONST_KEY in schema) {
     return schema.const;
   }
   throw new Error('schema cannot be inferred as a constant');
