+++ packages/utils/test/schema.test.ts
@@ -1,0 +1,24 @@
+import getTestValidator from './testUtils/getTestValidator';
+import {
+  getDefaultFormStateTest,
+  getDisplayLabelTest,
+  getMatchingOptionTest,
+  isFilesArrayTest,
+  isMultiSelectTest,
+  isSelectTest,
+  retrieveSchemaTest,
+  toIdSchemaTest,
+  toPathSchemaTest,
+} from './schema';
+
+const testValidator = getTestValidator({});
+
+getDefaultFormStateTest(testValidator);
+getDisplayLabelTest(testValidator);
+getMatchingOptionTest(testValidator);
+isFilesArrayTest(testValidator);
+isMultiSelectTest(testValidator);
+isSelectTest(testValidator);
+retrieveSchemaTest(testValidator);
+toIdSchemaTest(testValidator);
+toPathSchemaTest(testValidator);
